requirements.txt
# aiobotocore==2.22.0
# aiohappyeyeballs==2.6.1
# aiohttp==3.12.13
# aiohttp-retry==2.9.1
# aioitertools==0.12.0
# aiosignal==1.3.2
# albucore==0.0.24
# albumentations==2.0.8
# amqp==5.3.1
# annotated-types==0.7.0
# appdirs==1.4.4
# async-timeout==5.0.1
# asyncssh==2.21.0
# atpublic==6.0.1
# attrs==25.3.0
# augraphy==8.2.6
# billiard==4.2.1
# boto3==1.38.27
# botocore==1.38.27
# celery==5.5.3
# certifi==2025.6.15
# cffi==1.17.1
# charset-normalizer==3.4.2
# click==8.2.1
# click-didyoumean==0.3.1
# click-plugins==1.1.1.2
# click-repl==0.3.0
# colorama==0.4.6
# configobj==5.0.9
# contourpy==1.3.2
# cryptography==45.0.4
# cycler==0.12.1
# dictdiffer==0.9.0
# diskcache==5.6.3
# distro==1.9.0
# dpath==2.2.0
# dulwich==0.23.0
# dvc==3.60.1
# dvc-data==3.16.10
# dvc-http==2.32.0
# dvc-objects==5.1.1
# dvc-render==1.0.2
# dvc-s3==3.2.2
# dvc-studio-client==0.21.0
# dvc-task==0.40.2
# entrypoints==0.4
# filelock==3.18.0
# flatten-dict==0.4.2
# flufl.lock==8.2.0
# fonttools==4.58.4
# frozenlist==1.7.0
# fsspec==2025.5.1
# funcy==2.0
# gitdb==4.0.12
# GitPython==3.1.44
# grandalf==0.8
# gto==1.7.2
# hf-xet==1.1.5
# huggingface-hub==0.33.1
# hydra-core==1.4.0.dev1
# idna==3.10
# imageio==2.37.0
# importlib_metadata==8.7.0
# iterative-telemetry==0.0.10
# Jinja2==3.1.6
# jmespath==1.0.1
# joblib==1.5.1
# kiwisolver==1.4.8
# kombu==5.5.4
# lazy_loader==0.4
# lightning-utilities==0.14.3
# llvmlite==0.44.0
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# matplotlib==3.10.3
# mdurl==0.1.2
# mpmath==1.3.0
# multidict==6.6.2
# networkx==3.4.2
# numba==0.61.2
# numpy==2.2.6
# omegaconf==2.4.0.dev3
# opencv-python==4.11.0.86
# opencv-python-headless==4.11.0.86
# opentelemetry-api==1.34.1
# opentelemetry-sdk==1.34.1
# opentelemetry-semantic-conventions==0.55b1
# orjson==3.10.18
# packaging==25.0
# pandas==2.3.0
# pathspec==0.12.1
# pillow==11.2.1
# platformdirs==4.3.8
# prompt_toolkit==3.0.51
# propcache==0.3.2
# protobuf==6.31.1
# psutil==7.0.0
# pycparser==2.22
# pydantic==2.11.7
# pydantic_core==2.33.2
# pydot==4.0.1
# pygit2==1.18.0
# Pygments==2.19.2
# pygtrie==2.5.0
# pyparsing==3.2.3
# python-dateutil==2.9.0.post0
# pytorch-lightning==2.5.2
# pytz==2025.2
# PyYAML==6.0.2
# requests==2.32.4
# rich==14.0.0
# ruamel.yaml==0.18.14
# ruamel.yaml.clib==0.2.12
# s3fs==2025.5.1
# s3transfer==0.13.0
# safetensors==0.5.3
# scikit-image==0.25.2
# scikit-learn==1.7.0
# scipy==1.15.3
# scmrepo==3.3.11
# semver==3.0.4
# sentry-sdk==3.0.0a2
# shellingham==1.5.4
# shortuuid==1.0.13
# shtab==1.7.2
# simsimd==6.4.9
# six==1.17.0
# smmap==5.0.2
# sqltrie==0.11.2
# stringzilla==3.12.5
# sympy==1.14.0
# tabulate==0.9.0
# threadpoolctl==3.6.0
# tifffile==2025.5.10
# timm==1.0.16
# tomlkit==0.13.3
# torch==2.7.1
# torchmetrics==1.7.3
# torchvision==0.22.1
# tqdm==4.67.1
# typer==0.16.0
# typing-inspection==0.4.1
# typing_extensions==4.14.0
# tzdata==2025.2
# urllib3==2.5.0
# vine==5.1.0
# voluptuous==0.15.2
# wandb==0.20.2rc20250616
# wcwidth==0.2.13
# wrapt==1.17.2
# yarl==1.20.1
# zc.lockfile==3.0.post1
# zipp==3.23.0
# seaborn==0.13.2
# deskew

aiobotocore
aiohappyeyeballs
aiohttp
aiohttp-retry
aioitertools
aiosignal
albucore
albumentations
amqp
annotated-types
appdirs
async-timeout
asyncssh
atpublic
attrs
augraphy
billiard
boto3
botocore
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
configobj
contourpy
cryptography
cycler
dictdiffer
diskcache
distro
dpath
dulwich
dvc
dvc-data
dvc-http
dvc-objects
dvc-render
dvc-s3
dvc-studio-client
dvc-task
entrypoints
filelock
flatten-dict
flufl.lock
fonttools
frozenlist
fsspec
funcy
gitdb
GitPython
grandalf
gto
hf-xet
huggingface-hub
hydra-core
idna
imageio
importlib_metadata
iterative-telemetry
Jinja2
jmespath
joblib
kiwisolver
kombu
lazy_loader
lightning-utilities
llvmlite
markdown-it-py
MarkupSafe
matplotlib
mdurl
mpmath
multidict
networkx
numba
numpy
omegaconf
opencv-python
opencv-python-headless
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
packaging
pandas
pathspec
pillow
platformdirs
prompt_toolkit
propcache
protobuf
psutil
pycparser
pydantic
pydantic_core
pydot
pygit2
Pygments
pygtrie
pyparsing
python-dateutil
pytorch-lightning
pytz
PyYAML
requests
rich
ruamel.yaml
ruamel.yaml.clib
s3fs
s3transfer
safetensors
scikit-image
scikit-learn
scipy
scmrepo
semver
sentry-sdk
shellingham
shortuuid
shtab
# simsimd
six
smmap
sqltrie
stringzilla
sympy
tabulate
threadpoolctl
tifffile
timm
tomlkit
torch
torchmetrics
torchvision
tqdm
typer
# typing-inspection
typing_extensions
tzdata
urllib3
vine
voluptuous
wandb
wcwidth
wrapt
yarl
zc.lockfile
zipp
seaborn
deskew